requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.4.0
google-auth==2.25.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing-extensions==4.8.0
urllib3==2.1.0
werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0